+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Yii Event Change Log
 
-
 ## 1.0.2 under development
 
-- no changes in this release.
+- Enh #32: More detailed exception messages in `ListenerConfigurationChecker` (vjik)
 
 ## 1.0.1 August 30, 2021